requirements.txt
# Snowflake connectivity
snowflake-connector-python==3.12.2

# Azure services
azure-identity==1.18.0
azure-search-documents==11.6.0
openai==1.51.2

# Configuration and utilities
pyyaml==6.0.2
python-dotenv==1.0.1

# Data processing
pandas==2.2.3
numpy==2.1.2

# Logging and monitoring
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0

# Development tools
black==24.8.0
flake8==7.1.1
mypy==1.11.2
